requirements.txt
Flask
gunicorn
python-telegram-bot-raw